Renamed server to listener and listener to parser
